[refactor] CommentsInput과 CommentsList 컴포넌트에서 useUserNickname 훅을 엔트리 파일을 통해 불러오도록 수정
