+++ src/entities/auth/index.ts
@@ -1,6 +1,6 @@
 export * from './model/types';
 export * from './consts/authConstants';
-export {useIsLoggedIn, useUserEmail, useUserId, useUpdateUser} from './model/authStore';
+export {useIsLoggedIn, useUserNickname, useUpdateUser} from './model/authStore';
 export {getSession} from './apis/api-service';
 export {login, logout} from './apis/api-service';
 export {useGetSession} from './model/useGetSession';
